pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-content-ingestion-pipeline"
version = "0.1.0"
description = "A pipeline for ingesting documentation content, generating embeddings, and storing them in a vector database"
readme = "README.md"
authors = [{name = "Developer", email = "dev@example.com"}]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "cohere>=4.0.0",
    "qdrant-client>=1.9.0",
    "python-dotenv>=1.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0"
]

[project.scripts]
rag-ingest = "backend.src.cli.main_cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*"]